refactor(chat): move export/PDF code out of api/endpoints/apps.js; fix Markdown export dropping links/lists/headings
api/endpoints/apps.js was 914 lines but only ~130 were API calls; the rest was
PDF/HTML/JSON/Markdown export generation. Move that code into a dedicated
features/chat/utils/chatExport.js module so api/endpoints/* stays a thin HTTP
wrapper, and reuse the existing turndown-based isMarkdown/htmlToMarkdown from
utils/markdownUtils.js instead of a private regex converter that silently
dropped links, lists, and headings during Markdown export. Also deletes two
dead duplicate helpers (markdownToHtml, cleanHtmlForExport) that had zero call
sites and already exist in markdownUtils.js.

Fixes #1779
