+++ client/src/features/chat/components/ExportConversationMenu.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import Icon from '../../../shared/components/Icon';
-import { exportChatToFormat } from '../../../api/endpoints/apps';
+import { exportChatToFormat } from '../utils/chatExport';
 import { useUIConfig } from '../../../shared/contexts/UIConfigContext';
 import { getLocalizedContent } from '../../../utils/localizeContent';
 import useFeatureFlags from '../../../shared/hooks/useFeatureFlags';
